pipeline/requirements.txt
pylint
pytest
bs4
python-dotenv
requests
psycopg2-binary
flask
boto3
pytest-cov
flask_session
selenium
